+++ unkey/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Final
 
 __packagename__: Final[str] = "unkey.py"
-__version__: Final[str] = "0.3.0"
+__version__: Final[str] = "0.4.0"
 __author__: Final[str] = "Jonxslays"
 __copyright__: Final[str] = "2023-present Jonxslays"
 __description__: Final[str] = "An asynchronous Python SDK for unkey.dev."
@@ -21,6 +21,7 @@ from . import result
 from . import routes
 from . import serializer
 from . import services
+from . import undefined
 from .client import *
 from .errors import *
 from .models import *
@@ -28,6 +29,7 @@ from .result import *
 from .routes import *
 from .serializer import *
 from .services import *
+from .undefined import *
 
 __all__ = (
     "client",
@@ -38,6 +40,7 @@ __all__ = (
     "routes",
     "serializer",
     "services",
+    "undefined",
     "Api",
     "ApiKey",
     "ApiKeyList",
@@ -55,11 +58,15 @@ __all__ = (
     "HttpResponse",
     "HttpService",
     "KeyService",
+    "MissingRequiredArgument",
     "Ok",
     "Ratelimit",
     "RatelimitType",
     "Result",
     "Route",
     "Serializer",
+    "UndefinedNoneOr",
+    "UndefinedOr",
     "UnwrapError",
+    "UNDEFINED",
 )