+++ unkey/routes.py
@@ -83,6 +83,7 @@ class Route:
 CREATE_KEY: t.Final[Route] = Route(c.POST, "/keys")
 VERIFY_KEY: t.Final[Route] = Route(c.POST, "/keys/verify")
 REVOKE_KEY: t.Final[Route] = Route(c.DELETE, "/keys/{}")
+UPDATE_KEY: t.Final[Route] = Route(c.PUT, "/keys/{}")
 
 # Apis
 GET_API: t.Final[Route] = Route(c.GET, "/apis/{}")