+++ unkey/services/keys.py
@@ -2,9 +2,11 @@ from __future__ import annotations
 
 import typing as t
 
+from unkey import errors
 from unkey import models
 from unkey import result
 from unkey import routes
+from unkey import undefined
 
 from . import BaseService
 
@@ -25,15 +27,18 @@ class KeyService(BaseService):
         owner_id: str,
         prefix: str,
         *,
-        byte_length: t.Optional[int] = None,
-        meta: t.Optional[t.Dict[str, t.Any]] = None,
-        expires: t.Optional[int] = None,
-        remaining: t.Optional[int] = None,
-        ratelimit: t.Optional[models.Ratelimit] = None,
+        name: undefined.UndefinedOr[str] = undefined.UNDEFINED,
+        byte_length: undefined.UndefinedOr[int] = undefined.UNDEFINED,
+        meta: undefined.UndefinedOr[t.Dict[str, t.Any]] = undefined.UNDEFINED,
+        expires: undefined.UndefinedOr[int] = undefined.UNDEFINED,
+        remaining: undefined.UndefinedOr[int] = undefined.UNDEFINED,
+        ratelimit: undefined.UndefinedOr[models.Ratelimit] = undefined.UNDEFINED,
     ) -> ResultT[models.ApiKey]:
         """Creates a new api key.
 
         Args:
+            name: The name to use for this key.
+
             api_id: The id of the api this key is for.
 
             owner_id: The owner id to use for this key. Represents the
@@ -63,6 +68,7 @@ class KeyService(BaseService):
         route = routes.CREATE_KEY.compile()
         payload = self._generate_map(
             meta=meta,
+            name=name,
             apiId=api_id,
             prefix=prefix,
             ownerId=owner_id,
@@ -127,5 +133,64 @@ class KeyService(BaseService):
                     models.ErrorCode.from_str_maybe(data.get("code", "unknown")),
                 )
             )
+
+        return result.Ok(models.HttpResponse(200, "OK"))
+
+    async def update_key(
+        self,
+        key_id: str,
+        *,
+        name: undefined.UndefinedNoneOr[str] = undefined.UNDEFINED,
+        owner_id: undefined.UndefinedNoneOr[str] = undefined.UNDEFINED,
+        meta: undefined.UndefinedNoneOr[t.Dict[str, t.Any]] = undefined.UNDEFINED,
+        expires: undefined.UndefinedNoneOr[int] = undefined.UNDEFINED,
+        remaining: undefined.UndefinedNoneOr[int] = undefined.UNDEFINED,
+        ratelimit: undefined.UndefinedNoneOr[models.Ratelimit] = undefined.UNDEFINED,
+    ) -> ResultT[models.HttpResponse]:
+        """Updates an existing api key. To delete a key set its value
+        to `None`.
+
+        Args:
+            key_id: The id of the key to update.
+
+        Keyword Args:
+            name: The new name to use for this key.
+
+            owner_id: The new owner id to use for this key.
+
+            meta: The new dynamic mapping of information used
+                to provide context around this keys user.
+
+            expires: The new number of milliseconds into the future
+                when this key should expire.
+
+            remaining: The new max number of times this key can be
+                used.
+
+            ratelimit: The new Ratelimit to set on this key.
+
+        Returns:
+            A result containing the OK response or an error.
+        """
+        if undefined.all_undefined(name, owner_id, meta, expires, remaining, ratelimit):
+            raise errors.MissingRequiredArgument("At least one value is required to be updated.")
+
+        route = routes.UPDATE_KEY.compile(key_id)
+        payload = self._generate_map(
+            name=name,
+            meta=meta,
+            keyId=key_id,
+            ownerId=owner_id,
+            remaining=remaining,
+            ratelimit=ratelimit,
+            expires=self._expires_in(milliseconds=expires or 0)
+            if expires is not None
+            else expires,
+        )
+
+        data = await self._http.fetch(route, payload=payload)
+
+        if isinstance(data, models.HttpResponse):
+            return result.Err(data)
 
         return result.Ok(models.HttpResponse(200, "OK"))